refactor(app): FSD 구조의 app 레이어 형식에 맞게 파일 이동
 Provider 파일들을 app/provider로 이동 (ModalContext, QueryClientProvider, ReactQueryDevtools)
 타입 정의 파일들을 app/type으로 이동 (global.d.ts, svg.d.ts)
• Import 경로를 path-alias를 사용한 절대 경로로 변경
 FSD 아키텍처의 app 레이어 구조에 맞게 파일 정리
