+++ apps/web/src/app/type/global.d.ts
