+++ apps/web/src/app/type/svg.d.ts
